Improved overlay, Added color scale, Added interploation for numbers
Color scale will go from green to red as the player is closer to danager.

Numbers will move slowly over time to prevent instant spikes. Helps provide a much smoother transition and give an impression of movement.

Added remove timer for debug, might add a graphic later to replace it for users. As the information is useful but I don't want a raw tick value shown.