Adjusted cell render to connect to each other
Didn't notice this in the old code. Seems the models are designed to scale. I will need to remake these models to remove the need. As it makes the models look a little bad. As well z-fighting happens a bit on some parts. I remove as much z-fighting as possible. However, the bottom cell has z-fighting due to the model itself.