Toyed around with thermal math
Not entirely sure if this will work as expected. Not only have I not tested the code but am unsure if the math is correct. What is even more concerning is the unit scales. As kilo-joules might be too small a unit to store inside of a integer. Especially since some blocks will take millions of kilo-joules to change states.