Heat spread pathfinder works
Few minor oversights in implementing the pathfinder from memory. Not bad considering I rarely implement iterative pathfinders. From what I can see it is working though I have a feeling its not 100%. Main concern is that heat is not spreading to each tile correctly. This is due to the nearest neighbor check. Which checks all 6 sides for heat movement. Then it estimates how much heat would move to the current pathed tile. The problem is we only do this once per tile. So if a high value was set next to the tile after it ran. Then it would not get that heat value resulting in incorrect values. This is a minor issue as it shouldn't break the system. However, might result in cold spots on the map.