Smoothed rotate of turbine
Previous rotation was having a slight snapping when cycling back around to 360. Well not so much a snap but a slow down. Not entirely sure it was due to 360 wrap around or frame rate issues. Either way a lerp has been added and a synced rotation to tick rate to lerp towards.