Fixed fluid/item input for processing machines, Fixed boiler waste recipe
Created a system to check if any recipe supported a component. This is then used by inventory and fluid input to check if something should be allowed inside the machine.

Added to this fixed boiler trying to output 100 buckets of fluid of contaminated water. Small error due to configuration bug