Fixed map data unloading when a player dies
Is going to need more work as MC is unloading chunks on death. Which is clearing out the data (which it should do). However, the map is not designed to reload data  on its own. 

What is needed is to recycle the radiation sources on chunk data change (load, unload, blocks). This way the map state updates correctly.