Starting on thermal grid system
Going to replace how the map data structure works with a linked list like system. In this system each heat source will store a map of effect values. Then add that to the map.

The idea behind this is to ensure each heat source tracks its change. As well to provide an easy way to store the in progress heat maps.

With this new system (which is no where close to completed) we should be able to better track data. Now only per source but also per location. As any block edits can then map directly to source. Resulting in an easy way to update the game state. Without having to do complex checks to see which heat source needs to thread again.