Implemented wrench colorization and mode textures, Fixed layers overlapping
Wrench item layers had an issue with z-fighting for unknown reasons. Simple fix was applied that removed overlap in the layers. This should solve the problem but more research should be conducted to find the reason for z-fighting. Since in theory this should not happen as layers should merge down when backed... in theory.