Forgot to check sides for entry/exit test
Also neglected to make a note that no matter what it will detect entry/exit on a side even if it can't connect. To test for this and make it more clear I added a can connect check. This will later be used by the pathfinder and other systems to improve connection logic.